requirements.txt
Flask
numpy
tensorflow
pillow
gdown
gunicorn  # For deployment (optional)
